braille/requirements.txt
openai-whisper
sounddevice
scipy
pillow
tk